Remove category table date from redux state
This completely removes the data for the category tables from the redux
state, and instead gives the TaskItem container that responsibility. As
a result, this also makes it so that you no longer have to pass all of
the bid components to the select task action. Doing this also removes
the need to have the JobItem container since it no longer has to read
the redux state, as well as the category tables reducer.