Fix bug in array2Hash function (#39)
Each element of the hash returned by array2hash should be an array, but
the current implementation did not always do this. array2hash would
sometimes return an object as the value of an element if only a single
object from the original array matched the value for the specified key.