Rework fetches to use current bid (#42)
This changes the thunk chain to select a bid, then fetch those
routes, and use those routes for future fetches. This eliminates the
need for composing routes based on bid numbers so the bid argument from
these functions is removed. It also adds a fetchAllAndSelectFirst temporary
function which can be replaced when there is UI in place to choose a bid.