Add basic login
Add component and route for logging in. Django's login form is leveraged
for authentication since it bundles built in security. The default user
model is replaced by a custom model which is updated to link to the
Account model. Due to the order of migration application, when
initializing a fresh database, this may require migrating the 'bids' app
before the rest of the project.