Split dashboard into separate components
Separated the components that made up the dashboard (bidoverview and
bidselector) since they were no longer depended on each other. Deleted
the dashboard component since it was no longer being used. Refactored
the bidoverview and bidselector so that each component fetches it's own
data instead of depending on UI actions to fetch the necessary info.